gallerist/requirements.txt
azure-functions
gallerist
gallerist-azurestorage
essentials
roconfiguration
